update: move type alias of state management to own file
